+++ study/spring-1/src/main/java/tech/kingoyster/spring_1/customer/CustomerService.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface CustomerService {
     List<Customer> getAll();
 
-    Customer getById(Long id);
+    Customer getById(Integer id);
 
     Customer create(Customer customer);
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 }
